refactor(components): split to separate folders
