+++ frontend/app/components/admin/dashboard/index.ts
@@ -1,0 +1,2 @@
+export { default as AdminPanel } from './AdminPanel';
+export { default as QuizDetail } from './QuizDetail';
