+++ frontend/app/components/admin/editor/ImageDropzone.tsx
