+++ frontend/app/components/admin/index.ts
@@ -1,5 +1,4 @@
-export { default as AdminPanel } from './AdminPanel';
-export { default as QuizDetail } from './QuizDetail';
-export { default as QuizEditor } from './QuizEditor';
-export { default as Sidebar } from './Sidebar';
-export { default as TopBar } from './TopBar';
+export * from './dashboard';
+export * from './editor';
+export * from './layout';
+export * from './shared';
