+++ frontend/app/components/admin/layout/index.ts
@@ -1,0 +1,3 @@
+export { default as AdminLayout } from './AdminLayout';
+export { default as Sidebar } from './Sidebar';
+export { default as TopBar } from './TopBar';
