+++ frontend/app/components/quiz/index.ts
@@ -1,21 +1,3 @@
-export { default as AttemptReview } from './AttemptReview';
-export { default as CheckboxAnswer } from './CheckboxAnswer';
-export { default as ImageAnswers } from './ImageAnswers';
-export { default as ImageQuestion } from './ImageQuestion';
-export { default as MultipleChoice } from './MultipleChoice';
-export { default as Ordering } from './Ordering';
-export { default as ProgressBar } from './ProgressBar';
-export { default as QuestionCard } from './QuestionCard';
-export { default as QuestionHeader } from './QuestionHeader';
-export { default as QuizLayout } from './QuizLayout';
-export { default as QuizResults } from './QuizResults';
-export { default as QuizStart } from './QuizStart';
-export { default as RadioAnswer } from './RadioAnswer';
-export { default as RangeSlider } from './RangeSlider';
-export { default as RankingRow } from './RankingRow';
-export { default as ReviewQuestionCard } from './ReviewQuestionCard';
-export { default as ScoreCircle } from './ScoreCircle';
-export { default as SingleChoice } from './SingleChoice';
-export { default as SliderQuestion } from './SliderQuestion';
-export { default as SliderTrack } from './SliderTrack';
-export { default as SubmitButton } from './SubmitButton';
+export * from './questions';
+export * from './results';
+export * from './shared';
