+++ frontend/app/components/quiz/questions/index.ts
@@ -1,0 +1,7 @@
+export { default as ImageAnswers } from './ImageAnswers';
+export { default as ImageQuestion } from './ImageQuestion';
+export { default as MultipleChoice } from './MultipleChoice';
+export { default as Ordering } from './Ordering';
+export { default as RangeSlider } from './RangeSlider';
+export { default as SingleChoice } from './SingleChoice';
+export { default as SliderQuestion } from './SliderQuestion';
