+++ frontend/app/components/quiz/results/QuizResults.tsx
