+++ frontend/app/components/quiz/results/RankingRow.tsx
