+++ frontend/app/components/quiz/results/ScoreCircle.tsx
