+++ frontend/app/components/quiz/shared/CheckboxAnswer.tsx
