+++ frontend/app/components/quiz/shared/ProgressBar.tsx
