+++ frontend/app/components/quiz/shared/QuestionCard.tsx
