+++ frontend/app/components/quiz/shared/QuestionHeader.tsx
