+++ frontend/app/components/quiz/shared/QuizLayout.tsx
