+++ frontend/app/components/quiz/shared/QuizStart.tsx
