+++ frontend/app/components/quiz/shared/SliderTrack.tsx
